+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Fixed
+
+- CORS settings of the web_proxy now allow the token header to be provided.
+
 ## v2.17.0 - 2023-08-10
 
 ### Added